requirements.txt
torch
transformers
onnxruntime
gradio
pillow
numpy